pyproject.toml
[tool.poetry]
name = "ai-application"
version = "0.1.0"
description = ""
authors = ["Dajlan Zaimaj <51968936+dailani@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.66.2"
grpcio-tools = "^1.66.2"
openai = "^1.51.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
